requirements.txt
amqp==5.0.6
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
billiard==3.6.4.0
celery==5.1.2
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Django==3.2.8
django-js-asset==1.2.2
django-mptt==0.13.4
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
iniconfig==1.1.1
kombu==5.1.0
packaging==21.0
Pillow==8.3.2
pluggy==1.0.0
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
py==1.10.0
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.5
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
vine==5.0.0
wcwidth==0.2.5
